Moved URL tests to dedicated class, and relocated trailing slash logic to constructor
