+++ patient-generated-data/src/test/java/gov/va/api/health/patientgenerateddata/MetadataControllerTest.java
@@ -191,18 +191,4 @@ public class MetadataControllerTest {
                             .build()))
                 .build());
   }
-
-  @Test
-  void urlBuiltCorrectlyWithTrailingSlash() {
-    UrlPageLinks pageLinks =
-        UrlPageLinks.builder().baseUrl("http://va.gov/").r4BasePath("api/gov").build();
-    assertThat(pageLinks.r4Url()).isEqualTo("http://va.gov/api/gov");
-  }
-
-  @Test
-  void urlBuiltCorrectlyWithoutTrailingSlash() {
-    UrlPageLinks pageLinks =
-        UrlPageLinks.builder().baseUrl("http://va.gov").r4BasePath("api/gov").build();
-    assertThat(pageLinks.r4Url()).isEqualTo("http://va.gov/api/gov");
-  }
 }
